# Perf benchmark tests for span creation ( along with context management )