--- a/api/test/trace/CMakeLists.txt
+++ b/api/test/trace/CMakeLists.txt
@@ -24,3 +24,6 @@ endforeach()
 add_executable(span_id_benchmark span_id_benchmark.cc)
 target_link_libraries(span_id_benchmark benchmark::benchmark
                       ${CMAKE_THREAD_LIBS_INIT} opentelemetry_api)
+add_executable(span_benchmark span_benchmark.cc)
+target_link_libraries(span_benchmark benchmark::benchmark
+                      ${CMAKE_THREAD_LIBS_INIT} opentelemetry_api)