## Changes

As we are discussing performance changes for the api, it's important to get some numbers for span creation and context management involved in the process. The tests are:

- Test to measure performance for single-span creation
- Test to measure performance for single-span creation with scope ( involves stack )
- Test to measure performance for nested span creation with scope 
- Test to measure performance for nested span creation with manual span context management (no stack involved)
- Test to measure performance for nested span creation with context propagation

These are the numbers as tested locally:
```
------------------------------------------------------------------------------------------
Benchmark                                                Time             CPU   Iterations
------------------------------------------------------------------------------------------
BM_SpanCreation                                        263 ns          263 ns      2538457
BM_SpanCreationWithScope                              2222 ns         2221 ns       314466
BM_NestedSpanCreationWithScope                        7892 ns         7891 ns        87182
BM_SpanCreationWithManualSpanContextPropagation       1164 ns         1164 ns       629233
BM_SpanCreationWitContextPropagation                  6642 ns         6641 ns        97745
```

For significant contributions please make sure you have completed the following items:

* [ ] `CHANGELOG.md` updated for non-trivial changes
* [ ] Unit tests have been added
* [ ] Changes in public API reviewed